Backend/requirements.txt
Django==4.1.4
djangorestframework==3.14.0
geocoder==1.38.1
Requests==2.28.2
authentication==1.1.0
pipreqs==0.4.13
django-cors-headers==3.13.0
geocoder==1.38.1
psycopg2==2.9.5
gunicorn==20.1.0
whitenoise==6.4.0
Brotli==1.0.9